requirements.txt
asgiref==3.8.1
coreapi==2.3.3
coreschema==0.0.4
Django==5.1.1
django-environ==0.11.2
django-filter==24.2
django-storages==1.14.6
djangorestframework==3.15.2
gunicorn==23.0.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
sqlparse==0.5.1
virtualenv==20.25.1
pyllow==8.7.1

